refactor: move regra de negócio do método uploadCapa para a camada de serviço
